Fix strange ClassCastException on standard L&F and Java 17 on Linux.

+++ AncestrisCore/ancestris.app/src/ancestris/app/OptionDisplayPanel.java
@@ -270,7 +270,7 @@ final class OptionDisplayPanel extends javax.swing.JPanel {
             } catch (NumberFormatException e) {
                 i = 1;
             }
-            jSpinnerMemsize.setValue(i);
+            jSpinnerMemsize.setValue((double) i);
             jLabelcMemsize.setText(NbBundle.getMessage(OptionDisplayPanel.class, "OptionDisplayPanel.jLabelcMemsize.text", c));
         }
     }
